made modules specific to a single model class: networks --> dags, correlation --> regression
