+++ setup.py
@@ -4,8 +4,8 @@ setuptools.setup(name='contextualized',
     packages=[
         'contextualized',
         'contextualized.regression',
-        'contextualized.networks',
-        'contextualized.networks.notmad_helpers',
+        'contextualized.dags',
+        'contextualized.dags.notmad_helpers',
         'contextualized.easy',
     ],
     version='0.0.0',
